pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = ["robjsliwa <robjsliwa@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
mangum = "^0.17.0"
uvicorn = "^0.20.0"
fastapi = "^0.92.0"
mysql-connector-python = "^8.0.32"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
black = "^23.1.0"
flake8 = "^6.0.0"
httpx = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
